refactor: resolve compiler warnings

+++ include/point.h
@@ -20,7 +20,7 @@ std::string toString(const IntPoint& point,
 	const std::string& suffix = ")");
 
 std::string toString(const RealPoint& point,
-	size_t precision = 2,
+	int precision = 2,
 	const std::string& prefix = "(",
 	const std::string& separator = ", ",
 	const std::string& suffix = ")");
